+++ tests/integration/test_transaction.py
@@ -3251,6 +3251,21 @@ class TestTransaction(unittest.TestCase):
         self.assertEqual("3334445555", submitted_transaction.descriptor.phone)
         self.assertEqual("ebay.com", submitted_transaction.descriptor.url)
 
+    def test_submit_for_settlement_with_level_2_data(self):
+        transaction = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "4111111111111111",
+                "expiration_date": "05/2009"
+            }
+        }).transaction
+
+        params = {"purchase_order_number": "123456", "tax_amount": "2.00", "tax_exempt": False}
+
+        submitted_transaction = Transaction.submit_for_settlement(transaction.id, Decimal("900"), params).transaction
+
+        self.assertEqual(Transaction.Status.SubmittedForSettlement, submitted_transaction.status)
+
     @raises_with_regexp(KeyError, "'Invalid keys: invalid_param'")
     def test_submit_for_settlement_with_invalid_params(self):
         transaction = Transaction.sale({