+++ tests/integration/test_transaction.py
@@ -3266,6 +3266,32 @@ class TestTransaction(unittest.TestCase):
 
         self.assertEqual(Transaction.Status.SubmittedForSettlement, submitted_transaction.status)
 
+    def test_submit_for_settlement_with_level_3_data(self):
+        transaction = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "4111111111111111",
+                "expiration_date": "05/2009"
+            }
+        }).transaction
+
+        params = {
+                "discount_amount": "12.33",
+                "shipping_amount": "5.00",
+                "ships_from_postal_code": "90210",
+                "line_items": [{
+                    "quantity": "1.0232",
+                    "name": "Name #1",
+                    "kind": TransactionLineItem.Kind.Debit,
+                    "unit_amount": "45.1232",
+                    "total_amount": "45.15",
+                    }]
+                }
+
+        submitted_transaction = Transaction.submit_for_settlement(transaction.id, Decimal("900"), params).transaction
+
+        self.assertEqual(Transaction.Status.SubmittedForSettlement, submitted_transaction.status)
+
     @raises_with_regexp(KeyError, "'Invalid keys: invalid_param'")
     def test_submit_for_settlement_with_invalid_params(self):
         transaction = Transaction.sale({